Cast to UINT8
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk/trdp@1616 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ src/vos/windows/vos_thread.c
@@ -714,12 +714,12 @@ EXT_DECL void vos_getUuid (
     vos_getTime(&current);
 
     pUuID[0]    = current.tv_usec & 0xFF;
-    pUuID[1]    = (current.tv_usec & 0xFF00) >> 8;
-    pUuID[2]    = (current.tv_usec & 0xFF0000) >> 16;
-    pUuID[3]    = (current.tv_usec & 0xFF000000) >> 24;
+    pUuID[1]    = (UINT8)((current.tv_usec & 0xFF00) >> 8);
+    pUuID[2]    = (UINT8)((current.tv_usec & 0xFF0000) >> 16);
+    pUuID[3]    = (UINT8)((current.tv_usec & 0xFF000000) >> 24);
     pUuID[4]    = current.tv_sec & 0xFF;
     pUuID[5]    = (current.tv_sec & 0xFF00) >> 8;
-    pUuID[6]    = (current.tv_sec & 0xFF0000) >> 16;
+    pUuID[6]    = (UINT8)((current.tv_sec & 0xFF0000) >> 16);
     pUuID[7]    = ((current.tv_sec & 0x0F000000) >> 24) | 0x4; /*  pseudo-random version   */
 
     /* we always increment these values, this definitely makes the UUID unique */
